Private and public chat inbound now create trigger callbacks…with json arguments.
public-message
and
private-message

For both:
msginfo.nick = nickname
msginfo.body = body of the message

For public-message only:
msginfo.delayed = if the message inbound is an old queued message, this is flagged so it can be treated differently by the UI.
msginfo.action = if a message comes from the ROOM and not from an individual, this is set for UI purposes.
msginfo.nick_class = 'nick' if from someone else. and is = 'nick self' if it is a message which came from myself. All outbound messages wind up coming back as 'nick self' messages from xmpp servers.