Interim checkin. Content management is functional at the database level. Need to integrate with addspot/removespot/setspot and load-room and remove-room.
NOTE: dynamoDB rules for throughput will require a serious re-think on queuing of database commands so that we don't get killed on pricing/costs. Right now, we do detect when we are told that we have a throughput 'exceeds' but we don't re-try. We just call the failure callback. Not graceful yet.